Use BigInt for compiler representation of int literals
This is for future-proof: if we ever go to even bigger literals.

+++ stack-lang/js/JSCodeGen.scala
@@ -544,7 +544,7 @@ class JSCodeGen(runtime: JSRuntime, rewire: Map[Symbol, Symbol])(using defn: Def
     c match
       case Constant.Bool(b) => JS.BoolLit(b)
       case Constant.String(s) => JS.StringLit(s)
-      case Constant.Int(n) => JS.IntLit(n)
+      case Constant.Int(n) => JS.IntLit(n.toLong)
       case Constant.Float(d) => JS.FloatLit(d)
 
   /** Compile a list of expressions, correctly handling evaluation order */
